refactor: rename confix/config to config/index
