+++ apps/platform/src/app/components/homepage/HomepageIntro.tsx
@@ -5,7 +5,7 @@ import { useTranslations } from 'next-intl';
 import Container from '@/app/components/Container';
 import HomepageRoundBoxes from '@/app/components/homepage/HomepageRoundBoxes';
 
-import { CodaFormProjectLink } from '@/app/config/config';
+import { CodaFormProjectLink } from '@/app/config';
 
 type HomepageIntroProps = {
 	phases: RoundPhase[]
