+++ apps/platform/src/app/components/projects/ProjectsSidebar.tsx
@@ -4,7 +4,7 @@ import { useState } from 'react';
 import { useTranslations } from 'next-intl';
 import Image from 'next/image';
 
-import { neighborhoods } from '@/app/config/config';
+import { neighborhoods } from '@/app/config';
 import icons from '@/app/components/common/Icons';
 
 export default function ProjectsSidebar() {
@@ -19,7 +19,7 @@ export default function ProjectsSidebar() {
 		setSelectedNeighborhoods((prevSelected: string[]) =>
 			prevSelected.includes(neighborhood)
 				? prevSelected.filter((item: string) => item !== neighborhood)
-				: [...prevSelected, neighborhood], 
+				: [...prevSelected, neighborhood],
 		);
 	};
