+++ apps/platform/src/app/config/index.ts
